refactor(health): extract the pre-boot probe into a measured HealthProbe class
The SonarCloud gate failed PR #230 on new_coverage 0.0% against an 80% threshold, with every other condition green. Cause: api-sheriff/pom.xml excludes ApiSheriffApplication.class from JaCoCo — correct while that class was a bare Quarkus entry point, but this branch had added three testable methods to it, so the probe logic was measured nowhere while Sonar still counted its lines as new and uncovered.

Moving the probe into its own final class restores the exclusion's honesty rather than hollowing it out: the entry point is genuinely untestable (every path either exits the JVM or hands control to Quarkus) and stays excluded, while HealthProbe is measured and covered 7/7 lines by the existing tests. No probe behaviour changes.
